De-flexed my weatherbox. Ok kids: there is another widget toolkit made by adobe, it seems to be called the "v3 components". I wouldn't be surprised if the flex codebase grew out of these simple components.
To use them with the SDK, I had to create a SWC in the FAT containing
all these components (and their default skins). I am currently having
to compile this entire SWC into my swf so that I can get the skin classes
without having to enumerate them.

This checkin does not include that SWC I made.

These components are very basic: they don't resize if their contents are
too big, the scrolly list doesn't seem to understand the mouse wheel,
and the data handling is not as sophisticated as Flex's. Also, I really
liked setting up my UI in mxml, but now I'm having to hand-position
everything.

But they work. Easily. And even though I'm including all the components
and all the skins, the SWF is vastly smaller than the flex version.

More to the point: for most simple UI needs, like games, we should
*definitely* be using these components rather than the flex components.
The artists can drop them directly into the SWF assets, name them, and
we can populate them with data. We may even want to subclass these and
add our own functionality, but we'd want to then export the classes
back to the artists.

I'll be using these components in LOLcaptions, and having brittney
skin them and animate them, and it will be beautiful.