Ok, I actually want to divorce the default skins from the components because if the artists tweak a component skin in the FAT, it simply alters the "default". If we link-in the real defaults and load the artist SWF into the same ApplicationDomain then the defaults mask the artist customized skins.
So I'm creating a fl.skins package which contains references to the default
skin classes. If you want to use them, you simply need to use the class
Default*Skins class somewhere in your code.

I'm going to separate these for all the components, and then in the end I'll
create a SWC we can link against so that everyone can easily build projects
using these handy handy components.

Can we re-dist these components to creators?
Suuuuuuuure! It'll be fiiiine!