Removing the crash test, since it has served its purpose.
Crashes were due to an interaction of the FZip library, the "development" version of the
Linux Flash debug player (which came with the Flex 3 preview), and possibly the Flex 2
libraries. Switching from 9.0.63 alpha to 9.0.64 RC seems to have stopped crashes in this test
environment, although they still happen in the actual d2 project.

It looks like going back to the "official" 9.0.48 player is the only feasible solution for
now. We'll see if the next version of the player changes anything. Will pack loading completely
break when 9.0.64 RC becomes the official release version? Who knows, we'll find out.