Bringing over the remaining tuning parameters.
By the way, I love the way the Adobe XML parser handles syntax errors: it tells you that
the file failed to parse *in general*, but doesn't tell you what the problem was, or even
which line!

Because why should we coddle the end-user? They should consider themselves lucky they're
getting an XML parser at all! :)