After testing 3 different ways of rendering the board we discovered that, for a large board: - having a display object for every tree: 5fps - having a bitmap for a row of trees, and erasing bits for that row, which   allows us to still layer the trees: 10fps - going back to a flatter world, with no overlap between objects, so that   the trees and ground is all one big bitmap: 20fps.
So, we chose the fast way. It looks a little more "tiley", but I think
that's ok for this game. We still add factories as sprites, but we'll
be limiting the number of factories.

Also, bumped the tile size up to 31x31 from 29x29, but compensate by
only seeing half of the tile on the "horizon".