Removed the last bit of code to reference our path stuff. This uses HermiteFunc, none of the actual pathing.
Since I want to change HermiteFunc a bit at some point to
follow the standard easing function signature, this code would have to
change to adapt to that. Instead, I'm just going to not fuck around
and drag the current HermiteFunc in here for private use.
Zell can fix this up to use the standard hermite function if/when
we make it.