Frack! Updated to listen for the entity entered/left events so that I can remove the pointer for people who leave. It's a little wonky, because of course, this is a toy, so the entityIds of the avatars need to be correlated with the instance id of the viewer (memberId). Also, if a user changes their avatar while in the room, there is no event dispatched to indicate that the entityId!
Perhaps we should make the following changes:
- First of all, fix it so that changing avatars dispatches an
entityLeft/entityEntered combo with the new avatar entityId.
- Consider having memberEntered/memberLeft events so that
I can listen on those instead, as that's what I really care about.
The entityIds of their avatars don't matter so much to me.
- It might be useful to be able to look up the entityId of the
avatar of a member, given the memberId. I won't need this one if the
middle change is made, but it could still be very useful, rather than
having to iteratively ask each avatar if it's the one we're looking for.