Credits, and splash navigation to both credits and attempted levels.
"select_panel" goes on the splash screen after the first level is completed.  It lives in the lower left corner at 0, 391, and should have the level select and credits buttons (standard buttons) centered above each other on it.

For the credit screen, it should be based on black.  We show a smidgen of the Tesla level, placed at 152, 144 (no dashboard or any of that).  Populate it with dead units, and maybe have a single streetwalker wandering around aimlessly?
The actual credits display is in a movieclip called "credits" which goes at 350, 250.  Inside it are three placeholders with each of our first names.  We can put SW-facing units at those coordinates.

The text in the credits is going to need checking for PC/Mac font issues.