Upgraded Lineage:
We now only show a subset of the lineage to each player.  Each player sees their sire, and children+grandchilden.
Players now store their sireId and progeny ids.  That is how the Lineage is reconstructed on the server: offline player props are loaded as needed to show a player thier little Lineage.

This means that the database hit when a player starts will be fairly small, as very few players have a large horizontal lineage.
Also, the player experience should be much improved, as their lineage will never be partially there (missing players not currently online). It will upload to the client as the offline props are loaded.  So maybe the client sees a delay of a few seconds, but then their entire Lineage will be there.